LibreNMS library and __init__ fusion
